+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # CHANGELOG
 
+* v0.0.2 2019-8-24
+
+  ## 如何选择一个合适的 `deep copy default` 库
+  * 需求
+    
+    当源数据中的属性值(需要递归嵌套对象)为 `null` 或者 `undefined` 时, 使用默认值来对齐
+  
+    ```
+    结果                         源数据                   默认值
+    {                            {                       {
+        a: 'default-a',              a: null,                a: 'default-a',
+        b: {                         b: {                    b: {
+            bb: 'source',     <─         bb: 'source' <─         bb: 'default-bb',
+           bb1: 'default-bb1'                                   bb1: 'default-bb1'
+        }                            }                       }
+    }                            }                       }
+    ```
+
+    ```javascript
+    var a = deepCopyDefault({
+        a: null,
+        b: {
+            bb: 'source'
+        }
+    }, {
+        a: 'default-a',
+        b: {
+            bb: 'default-bb',
+            bb1: 'default-bb1'
+        }
+    });
+    // {a: 'default-a', b: {bb: 'source', bb1: 'default-bb1'}}
+    ```
+  * 首选
+    * https://github.com/sttk/fav-prop.defaults-deep
+  * 备选
+    * [defaults-deep](https://github.com/jonschlinkert/defaults-deep) 依赖较多
+  * 不合适
+    * [lodash.defaultsdeep](https://github.com/lodash/lodash/) `object` 中的属性值为 `null` 时没有从 `source` 中补齐
+    * [defaulty](https://github.com/fabioricali/defaulty) `target` 中的属性值为 `null` 时报错
+    * [default-assign](https://github.com/nanowizard/default-assign) `target` 中的属性值为 `null` 时没有从 `source` 中补齐, 只有当 `target` 中的属性值为 `undefined` 时才有效
+    * [object.defaults](https://github.com/jonschlinkert/object.defaults) 不支持 `deep`
+    * [defaults](https://github.com/jonschlinkert/object.defaults) 不支持 `deep`
+
 * v0.0.1 2019-8-8
 
   * 初始版本